Rename training route file and data
